Move IntermediateTaskGroups and its storage to src/Query/
